requirements.txt
ariadne==0.12.0
astroid==2.5.3
attrs==20.3.0
click==7.1.2
coverage==5.5
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
graphql-core==3.0.5
importlib-metadata==3.10.0
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pylint==2.7.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.15.0
SQLAlchemy==1.3.23
starlette==0.13.8
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
